Add support for getting a single episode from `nielsen tv info`
